--- Merging r34463 into '.': U    packages/fcl-base/src/inifiles.pp --- Recording mergeinfo for merge of r34463 into '.':  U   . --- Merging r34477 into '.': A    packages/paszlib/tests/tczstreamseek.pp U    packages/paszlib/src/zstream.pp --- Recording mergeinfo for merge of r34477 into '.':  G   . --- Merging r35100 into '.': U    rtl/freebsd/i386/identpatch.sh --- Recording mergeinfo for merge of r35100 into '.':  G   . --- Merging r35193 into '.': U    rtl/gba/sysutils.pp U    rtl/nds/sysutils.pp --- Recording mergeinfo for merge of r35193 into '.':  G   .
# revisions: 34463,34477,35100,35193

git-svn-id: branches/fixes_3_0@35306 -

+++ rtl/freebsd/i386/identpatch.sh
@@ -1,12 +1,18 @@
 #!/bin/sh
+if [ "$#" -ne 1 ]
+then
 elfdump -n `which elfdump` |awk '/FreeBSD/{print $2}' >elfversion
 IDVERSION=`cat elfversion`
 rm elfversion
+else
+IDVERSION=$1
+fi
 echo Patching cprt0.as with version $IDVERSION
 
 sed -I.sav -es/900044/$IDVERSION/ cprt0.as
 sed -I.sav -es/900044/$IDVERSION/ dllprt0.as
 sed -I.sav -es/900044/$IDVERSION/ prt0.as
+sed -I.sav -es/900044/$IDVERSION/ gprt0.as
 sed -I.sav -es/900044/$IDVERSION/ si_c.inc
 sed -I.sav -es/900044/$IDVERSION/ si_prc.inc
